+++ learn-designpatterns-in-csharp/Source/Behavioral/Mediator/ChatRoomDemo/ChatRoomDemo.csproj
@@ -7,8 +7,4 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
-  <ItemGroup>
-    <Folder Include="Mediators\" />
-  </ItemGroup>
-
 </Project>
