+++ learn-designpatterns-in-csharp/Source/Behavioral/Mediator/ChatRoomDemo/Mediators/IChatRoom.cs
@@ -17,44 +17,4 @@ namespace ChatRoomDemo.Mediators
         void SendTo<T>(string from, string message) where T : TeamMember;
     }
 
-
-    /// <summary>
-    /// ConcreteMediator
-    /// </summary>
-    public class TeamChatRoom : IChatRoom
-    {
-        private readonly Dictionary<string, TeamMember> teamMembers = new();
-
-        public void Register(TeamMember teamMember)
-        {
-            teamMember.SetChatroom(this);
-            if (!teamMembers.ContainsKey(teamMember.Name))
-            {
-                teamMembers.Add(teamMember.Name, teamMember);
-            }
-        }
-
-        public void Send(string from, string message)
-        {
-            foreach (var teamMember in teamMembers.Values)
-            {
-                teamMember.Receive(from, message);
-            }
-        }
-
-        public void Send(string from, string to, string message)
-        {
-            var teamMember = teamMembers[to];
-            teamMember?.Receive(from, message);
-        }
-
-        public void SendTo<T>(string from, string message) where T : TeamMember
-        {
-            foreach (var teamMember in teamMembers.Values.OfType<T>())
-            {
-                teamMember.Receive(from, message);
-            }
-        }
-    }
-
 }
